Added weather.py.
weather.py detects current weather conditions using Yahoo's! weather
API. The user must supply a 7 digit area code in their request, which is
automatically detected and extracted using a regular expression. As of
this commit, weather.py outputs the current weather conditions as well
as a forecast for the next week.
It can be invoked using the `hey` command as follows:
`hey tell me how the weather is looking for [7 digit code] please`