Added blend.py for the blending of colors.
blend.py accepts an arbitrary number of colors in hexadecimal notation.
It then proceeds to 'blend' (average) the hexadecimal numbers by
converting every 8 byte portion of 2 symbols to an integer, adding them
up respective to red, green and blue and then dividing them by however
many numbers were passed in. blend.py then outputs the blended color in
hexadecimal notation, as well as offering to open the color in a web
browser for the user to see.